requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
dominate==2.6.0
Flask==1.0.3
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
greenlet==1.1.0
gunicorn==20.1.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
python-dotenv==0.17.1
requests==2.22.0
SQLAlchemy==1.4.15
stripe==2.29.3
toml==0.9.6
urllib3==1.25.3
visitor==0.1.3
Werkzeug==0.15.5
WTForms==2.3.3
